requirements.txt
discord.py
yt_dlp
python-dotenv
PyNaCl
flask
beautifulsoup4
requests